requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
Faker==37.5.3
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
requests==2.32.4
tzdata==2025.2
urllib3==2.5.0
